Create GetAllDocumentIdsCallable from getAllDocumentIds method
This callable was standalone so can be pulled out with very
few changes -- literally just the IDEA refactor works for this
one, with a class-comment as to the class's purpose.
